bots/WORKER-rightsem-final/requirements.txt
cassandra-driver
ipwhois
dnspython
